+++ shelley/chain-and-ledger/executable-spec/shelley-spec-ledger.cabal
@@ -40,6 +40,7 @@ library
                      Shelley.Spec.Ledger.TxData
                      Shelley.Spec.Ledger.Updates
                      Shelley.Spec.Ledger.Validation
+                     Shelley.Spec.Ledger.Scripts
                      Shelley.Spec.Ledger.STS.Avup
                      Shelley.Spec.Ledger.STS.Bbody
                      Shelley.Spec.Ledger.STS.Tick
