+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/Tx.hs
@@ -33,13 +33,11 @@ where
 import           Shelley.Spec.Ledger.BaseTypes (invalidKey)
 import           Shelley.Spec.Ledger.Keys (AnyKeyHash, GenKeyHash, undiscriminateKeyHash)
 
-import           Cardano.Binary (FromCBOR (fromCBOR), ToCBOR (toCBOR), decodeListLenOf, decodeWord,
-                     encodeListLen, encodeMapLen, encodeWord, encodeWord8)
-import           Cardano.Crypto.Hash (hashWithSerialiser)
+import           Cardano.Binary (FromCBOR (fromCBOR), ToCBOR (toCBOR), decodeWord,
+                     encodeListLen, encodeWord, encodeMapLen, decodeListLenOf)
 import           Cardano.Ledger.Shelley.Crypto
 import           Cardano.Prelude (NoUnexpectedThunks (..), catMaybes)
 import           Data.Foldable (fold, toList)
-import qualified Data.List as List (concat, concatMap, permutations)
 import           Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import           Data.Maybe (mapMaybe)
@@ -50,10 +48,10 @@ import           GHC.Generics (Generic)
 import           Shelley.Spec.Ledger.MetaData (MetaData)
 
 import           Shelley.Spec.Ledger.Serialization (CborSeq (..), decodeMapContents)
-import           Shelley.Spec.Ledger.TxData (Credential (..), MultiSig (..), Script (..),
-                     ScriptHash (..), TxBody (..), TxId (..), TxIn (..), TxOut (..), WitVKey (..),
-                     nativeMultiSigTag, witKeyHash)
-
+import           Shelley.Spec.Ledger.TxData (Credential (..),
+                     TxBody (..), TxId (..), TxIn (..), TxOut (..), WitVKey (..),
+                     witKeyHash)
+import           Shelley.Spec.Ledger.Scripts
 
 -- |A fully formed transaction.
 data Tx crypto
@@ -140,13 +138,18 @@ instance Crypto crypto => FromCBOR (Tx crypto) where
   fromCBOR = decodeListLenOf 3 >>
     cborWitsToTx <$> fromCBOR <*> fromCBOR <*> fromCBOR
 
-
 -- | Typeclass for multis-signature script data types. Allows for script
 -- validation and hashing.
 class (Crypto crypto, ToCBOR a) =>
   MultiSignatureScript a crypto where
   validateScript :: a -> Tx crypto -> Bool
   hashScript :: a -> ScriptHash crypto
+
+-- | instance of MultiSignatureScript type class
+instance Crypto crypto =>
+  MultiSignatureScript (MultiSig crypto) crypto where
+  validateScript = validateNativeMultiSigScript
+  hashScript = \x -> hashAnyScript (MultiSigScript x)
 
 -- | Script evaluator for native multi-signature scheme. 'vhks' is the set of
 -- key hashes that signed the transaction to be validated.
@@ -173,50 +176,6 @@ validateNativeMultiSigScript msig tx =
   where witsSet = _witnessVKeySet tx
         vhks    = Set.map witKeyHash witsSet
 
--- | Hashes native multi-signature script, appending the 'nativeMultiSigTag' in
--- front and then calling the script CBOR function.
-hashAnyScript
-  :: Crypto crypto
-  => Script crypto
-  -> ScriptHash crypto
-hashAnyScript (MultiSigScript msig) =
-  ScriptHash $ hashWithSerialiser (\x -> encodeWord8 nativeMultiSigTag
-                                          <> toCBOR x) (MultiSigScript msig)
-
--- | Get one possible combination of keys for multi signature script
-getKeyCombination :: MultiSig crypto -> [AnyKeyHash crypto]
-
-getKeyCombination (RequireSignature hk) = [hk]
-getKeyCombination (RequireAllOf msigs) =
-  List.concatMap getKeyCombination msigs
-getKeyCombination (RequireAnyOf msigs) =
-  case msigs of
-    []  -> []
-    x:_ -> getKeyCombination x
-getKeyCombination (RequireMOf m msigs) =
-  List.concatMap getKeyCombination (take m msigs)
-
-
--- | Get all valid combinations of keys for given multi signature. This is
--- mainly useful for testing.
-getKeyCombinations :: MultiSig crypto -> [[AnyKeyHash crypto]]
-
-getKeyCombinations (RequireSignature hk) = [[hk]]
-
-getKeyCombinations (RequireAllOf msigs) = [List.concat $
-  List.concatMap getKeyCombinations msigs]
-
-getKeyCombinations (RequireAnyOf msigs) = List.concatMap getKeyCombinations msigs
-
-getKeyCombinations (RequireMOf m msigs) =
-  let perms = map (take m) $ List.permutations msigs in
-    map (concat . List.concatMap getKeyCombinations) perms
-
-
-instance Crypto crypto =>
-  MultiSignatureScript (MultiSig crypto) crypto where
-  validateScript = validateNativeMultiSigScript
-  hashScript = \x -> hashAnyScript (MultiSigScript x)
 
 -- | Multi-signature script witness accessor function for Transactions
 txwitsScript
