+++ shelley/chain-and-ledger/executable-spec/src/Shelley/Spec/Ledger/TxData.hs
@@ -38,15 +38,16 @@ import           Numeric.Natural (Natural)
 import           Ledger.Core (Relation (..))
 import           Shelley.Spec.Ledger.BaseTypes (Text64, UnitInterval, invalidKey)
 import           Shelley.Spec.Ledger.Coin (Coin (..))
-import           Shelley.Spec.Ledger.Keys (AnyKeyHash, pattern AnyKeyHash, GenKeyHash, Hash,
+import           Shelley.Spec.Ledger.Keys (AnyKeyHash, GenKeyHash, Hash,
                      KeyHash, pattern KeyHash, Sig, VKey, VerKeyVRF, hashAnyKey)
 import           Shelley.Spec.Ledger.MetaData (MetaDataHash)
 import           Shelley.Spec.Ledger.Slot (EpochNo (..), SlotNo (..))
 import           Shelley.Spec.Ledger.Updates (Update, emptyUpdate, updateNull)
 
 import           Shelley.Spec.Ledger.Serialization (CBORGroup (..), CBORMap (..), CborSeq (..),
-                     FromCBORGroup (..), ToCBORGroup (..), decodeList, decodeMapContents,
-                     encodeFoldable)
+                     FromCBORGroup (..), ToCBORGroup (..), decodeMapContents)
+
+import           Shelley.Spec.Ledger.Scripts
 
 -- |The delegation of one stake key to another.
 data Delegation crypto = Delegation
@@ -128,63 +129,6 @@ data Ptr
   deriving (ToCBOR, FromCBOR) via CBORGroup Ptr
 
 instance NoUnexpectedThunks Ptr
-
--- | A simple language for expressing conditions under which it is valid to
--- withdraw from a normal UTxO payment address or to use a stake address.
---
--- The use case is for expressing multi-signature payment addresses and
--- multi-signature stake addresses. These can be combined arbitrarily using
--- logical operations:
---
--- * multi-way \"and\";
--- * multi-way \"or\";
--- * multi-way \"N of M\".
---
--- This makes it easy to express multi-signature addresses, and provides an
--- extension point to express other validity conditions, e.g., as needed for
--- locking funds used with lightning.
---
-data MultiSig crypto =
-       -- | Require the redeeming transaction be witnessed by the spending key
-       --   corresponding to the given verification key hash.
-       RequireSignature   (AnyKeyHash crypto)
-
-       -- | Require all the sub-terms to be satisfied.
-     | RequireAllOf      [MultiSig crypto]
-
-       -- | Require any one of the sub-terms to be satisfied.
-     | RequireAnyOf      [MultiSig crypto]
-
-       -- | Require M of the given sub-terms to be satisfied.
-     | RequireMOf    Int [MultiSig crypto]
-  deriving (Show, Eq, Ord, Generic)
-
-instance NoUnexpectedThunks (MultiSig crypto)
-
--- | Magic number representing the tag of the native multi-signature script
--- language. For each script language included, a new tag is chosen and the tag
--- is included in the script hash for a script.
-nativeMultiSigTag :: Word8
-nativeMultiSigTag = 0
-
-newtype ScriptHash crypto =
-  ScriptHash (Hash (HASH crypto) (Script crypto))
-  deriving (Show, Eq, Ord, NoUnexpectedThunks)
-
-data Script crypto = MultiSigScript (MultiSig crypto)
-                  -- constructors for new languages go here
-                   -- e.g | PlutusScriptV1 ScriptPLC
-  deriving (Show, Eq, Ord, Generic)
-
-deriving instance Crypto crypto => ToCBOR (ScriptHash crypto)
-deriving instance Crypto crypto => FromCBOR (ScriptHash crypto)
-
--- | Count nodes and leaves of multi signature script
-countMSigNodes :: MultiSig crypto -> Int
-countMSigNodes (RequireSignature _) = 1
-countMSigNodes (RequireAllOf msigs) = 1 + sum (map countMSigNodes msigs)
-countMSigNodes (RequireAnyOf msigs) = 1 + sum (map countMSigNodes msigs)
-countMSigNodes (RequireMOf _ msigs) = 1 + sum (map countMSigNodes msigs)
 
 newtype Wdrl crypto = Wdrl { unWdrl :: Map (RewardAcnt crypto) Coin }
   deriving (Show, Eq, Generic, NoUnexpectedThunks)
@@ -522,43 +466,6 @@ instance
           , _mdHash   = Nothing
           }
 
-instance (Crypto crypto) =>
-  ToCBOR (MultiSig crypto) where
-  toCBOR (RequireSignature hk) =
-    encodeListLen 2 <> encodeWord 0 <> toCBOR hk
-  toCBOR (RequireAllOf msigs) =
-    encodeListLen 2 <> encodeWord 1 <> encodeFoldable msigs
-  toCBOR (RequireAnyOf msigs) =
-    encodeListLen 2 <> encodeWord 2 <> encodeFoldable msigs
-  toCBOR (RequireMOf m msigs) =
-    encodeListLen 3 <> encodeWord 3 <> toCBOR m <> encodeFoldable msigs
-
-instance (Crypto crypto) =>
-  FromCBOR (MultiSig crypto) where
-  fromCBOR = do
-    n <- decodeListLen
-    decodeWord >>= \case
-      0 -> matchSize "RequireSignature" 2 n >> (RequireSignature . AnyKeyHash) <$> fromCBOR
-      1 -> matchSize "RequireAllOf" 2 n >> RequireAllOf <$> decodeList fromCBOR
-      2 -> matchSize "RequireAnyOf" 2 n >> RequireAnyOf <$> decodeList fromCBOR
-      3 -> do
-        matchSize "RequireMOf" 3 n
-        m     <- fromCBOR
-        msigs <- decodeList fromCBOR
-        pure $ RequireMOf m msigs
-      k -> invalidKey k
-
-instance (Crypto crypto) =>
-  ToCBOR (Script crypto) where
-  toCBOR (MultiSigScript msig) =
-    toCBOR nativeMultiSigTag <> toCBOR msig
-
-instance (Crypto crypto) =>
-  FromCBOR (Script crypto) where
-  fromCBOR = do
-    decodeWord >>= \case
-      0 -> MultiSigScript <$> fromCBOR
-      k -> invalidKey k
 
 instance (Typeable crypto, Crypto crypto)
   => ToCBORGroup (Credential crypto) where
