+++ shelley/chain-and-ledger/executable-spec/test/Test/Shelley/Spec/Ledger/Shrinkers.hs
@@ -14,7 +14,7 @@ import           Shelley.Spec.Ledger.Slot
 import           Shelley.Spec.Ledger.Tx
 import           Shelley.Spec.Ledger.TxData
 import           Shelley.Spec.Ledger.Updates
-
+import           Shelley.Spec.Ledger.Scripts 
 import           Test.Shelley.Spec.Ledger.ConcreteCryptoTypes (Block)
 
 shrinkBlock
